+++ client/components/account-page.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import { connect } from 'react-redux'
 import { Link } from 'react-router-dom'
 
 export const AccountPage = () => {
@@ -16,12 +15,4 @@ export const AccountPage = () => {
     )
 }
 
-const mapState = (state) => {
-
-    return {
-        wineListOnProps: state.wine.wineList
-    }
-}
-
-
-export default connect(mapState)(AccountPage)
+export default AccountPage;
