+++ client/components/cart.js
@@ -4,7 +4,6 @@ import { Link } from 'react-router-dom'
 import { fetchItems, putItems, updateQuantity, removeItem } from '../store'
 
 
-
 export class Cart extends React.Component {
     constructor(props) {
         super(props)
