+++ client/components/checkout.js
@@ -81,4 +81,3 @@ const mapDispatch = (dispatch) => {
 
 
 export default connect(mapState, mapDispatch)(Checkout)
-
