+++ client/components/order-details.js
@@ -44,8 +44,6 @@ class OrderDetails extends React.Component {
 }
 
 
-
-
 const mapState = (state) => {
   console.log('state: ', state);
   return {
