+++ client/components/user-place-form.js
@@ -94,7 +94,6 @@ export const UserPlace = (props) => {
 }
 
 
-
 const mapState = (state) => {
     return {
         loggedInUser: state.user.loggedInUser
@@ -128,54 +127,3 @@ const mapDispatch = (dispatch, ownProps) => ({
 })
 
 export default connect(mapState, mapDispatch)(UserPlace)
-
-
-
-// <div>
-// <label>Address:</label>
-// <input
-//     name="address"
-//     type="text"
-//     defaultValue={user.place.address}
-// />
-// </div>
-// <div>
-// <label>City:</label>
-// <input
-//     name="city"
-//     type="text"
-//     defaultValue={user.place.city}
-// />
-// </div>
-// <div>
-// <label>State:</label>
-// <input
-//     name="state"
-//     type="text"
-//     defaultValue={user.place.state}
-// />
-// </div>
-// <div>
-//     <label>Zip Code:</label>
-//     <input
-//         name="zipcode"
-//         type="number"
-//         defaultValue={user.place.zipcode}
-//     />
-// </div>
-// <div>
-//     <label>Country:</label>
-//     <input
-//         name="country"
-//         type="text"
-//         defaultValue={user.place.country}
-//     />
-// </div>
-// <div>
-//     <label>Phone:</label>
-//     <input
-//         name="phone"
-//         type="tel"
-//         defaultValue={user.place.phone}
-//     />
-// </div>
